+++ app/src/androidTest/java/net/skyscanner/backpack/compose/BpkTextTest.kt
@@ -2,12 +2,12 @@ package net.skyscanner.backpack.compose
 
 import androidx.compose.material.Text
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import net.skyscanner.backpack.BpkComposeTest
+import net.skyscanner.backpack.BpkSnapshotTest
 import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class BpkTextTest : BpkComposeTest {
+class BpkTextTest : BpkSnapshotTest() {
 
   @Test
   fun default() = composed {
